Bugfix delete issue

+++ src/api/coworkers/delete.ts
@@ -14,12 +14,15 @@ export const del = (app, db) => {
                 .where({id})
 
             const coworkerResult = await coworker;
-            const userId: number = coworkerResult[0].user_id;
 
-            // Delete User
-            await db.delete().table("coworkers_plans").where({coworker_id: id})
-            await db.delete().table("coworkers").where({id})
-            await db.delete().table("users").where({id: userId})
+            if (coworkerResult) {
+                const userId: number = coworkerResult[0].user_id;
+
+                // Delete User
+                await db.delete().table("coworkers_plans").where({coworker_id: id})
+                await db.delete().table("coworkers").where({id})
+                await db.delete().table("users").where({id: userId})
+            }
 
             res.json({ message: 'Success' });
         } catch (err) {
